refactor(renderer): extract common logic from Electron and Tauri connectors
