+++ app/renderer/src/App.tsx
@@ -1,4 +1,4 @@
-import React, { Suspense, useEffect } from "react";
+import React, { Suspense } from "react";
 import { HashRouter as Router, Switch, Route } from "react-router-dom";
 import {
   ThemeProvider,
@@ -9,41 +9,14 @@ import { Layout, Preloader } from "components";
 import { compactRoutes, routes } from "config";
 import { useSelector } from "react-redux";
 import { AppStateTypes } from "store";
+import { useDisableNonNativeShortcuts } from "hooks/useDisableNonNativeShortcuts";
 
 export default function App() {
   const settings = useSelector(
     (state: AppStateTypes) => state.settings
   );
 
-  useEffect(() => {
-    const contextEvent = (event: MouseEvent) => {
-      if (event.target) {
-        let target = event.target as HTMLElement;
-        if (
-          target.tagName === "TEXTAREA" ||
-          (target.tagName === "INPUT" &&
-            (target as HTMLInputElement).type === "text")
-        ) {
-          return true;
-        }
-      }
-      event.preventDefault();
-      return false;
-    };
-    document.addEventListener("contextmenu", contextEvent);
-    return () =>
-      document.removeEventListener("contextmenu", contextEvent);
-  }, []);
-
-  useEffect(() => {
-    const middleMouseEvent = (event: MouseEvent) => {
-      if (event.button === 1) event.preventDefault();
-    };
-    window.addEventListener("auxclick", middleMouseEvent);
-
-    return () =>
-      window.removeEventListener("auxclick", middleMouseEvent);
-  }, []);
+  useDisableNonNativeShortcuts();
 
   return (
     <ThemeProvider>
