--- app/renderer/src/contexts/InvokeConnector.tsx
@@ -1,6 +1,0 @@
-/**
- * Explicitly for calling invokes from the trigger rather than a setting change.
- */
-export type InvokeConnector = {
-  send: (event: string, ...payload: any) => void;
-};
