+++ app/renderer/src/contexts/connectors/TauriConnector.tsx
@@ -1,19 +1,11 @@
-import React, { useCallback, useContext, useEffect } from "react";
+import React, { useCallback, useEffect } from "react";
 import { ConnnectorContext } from "../ConnnectorContext";
 import { useDispatch, useSelector } from "react-redux";
 import { AppStateTypes, SettingTypes } from "../../store";
-import { CounterContext } from "../CounterContext";
 import {
   CHECK_FOR_UPDATES,
-  SET_ALWAYS_ON_TOP,
   CLOSE_WINDOW,
-  SET_COMPACT_MODE,
-  SET_FULLSCREEN_BREAK,
   MINIMIZE_WINDOW,
-  SET_NATIVE_TITLEBAR,
-  SHOW_WINDOW,
-  SET_UI_THEME,
-  TRAY_ICON_UPDATE,
   UPDATE_AVAILABLE,
 } from "@pomatez/shareables";
 import {
@@ -21,19 +13,11 @@ import {
   disable,
   isEnabled,
 } from "@tauri-apps/plugin-autostart";
-import { invoke } from "@tauri-apps/api/primitives";
+
 import { listen } from "@tauri-apps/api/event";
 import { open } from "@tauri-apps/plugin-shell";
 import { setUpdateBody, setUpdateVersion } from "../../store/update";
-import { useTrayIconUpdates } from "hooks/useTrayIconUpdates";
-
-export const TauriInvokeConnector = {
-  send: (event: string, ...payload: any) => {
-    invoke(event.toLowerCase(), ...payload).catch((err) =>
-      console.error(err)
-    );
-  },
-};
+import { TauriInvoker } from "contexts/InvokeConnectors";
 
 export const TauriConnectorProvider: React.FC = ({ children }) => {
   const dispatch = useDispatch();
@@ -83,15 +67,6 @@ export const TauriConnectorProvider: React.FC = ({ children }) => {
     };
   }, []);
 
-  /**
-   * Rust uses lowercase snake_case for function names, so we need to convert to lower case for the calls.
-   * @param event
-   * @param payload
-   */
-  const send = useCallback(async (event: string, ...payload: any) => {
-    await invoke(event.toLowerCase(), ...payload);
-  }, []);
-
   useEffect(() => {
     // The autostart-plugin fails when trying to disble if it is already disabled
     // https://github.com/tauri-apps/plugins-workspace/issues/24#issuecomment-1528958008
@@ -114,75 +89,30 @@ export const TauriConnectorProvider: React.FC = ({ children }) => {
     });
   }, [settings.openAtLogin]);
 
-  const timer = useSelector((state: AppStateTypes) => state.timer);
-
-  const { shouldFullscreen } = useContext(CounterContext);
-
   const onMinimizeCallback = useCallback(() => {
-    send(MINIMIZE_WINDOW, {
+    TauriInvoker.send(MINIMIZE_WINDOW, {
       minimizeToTray: settings.minimizeToTray,
     });
     console.log("Minimize callback");
-  }, [send, settings.minimizeToTray]);
+  }, [settings.minimizeToTray]);
 
   const onExitCallback = useCallback(() => {
-    send(CLOSE_WINDOW, {
+    TauriInvoker.send(CLOSE_WINDOW, {
       closeToTray: settings.closeToTray,
     });
-  }, [send, settings.closeToTray]);
+  }, [settings.closeToTray]);
 
   /**
    * Not needed as tauri already opens these externally.
    */
   const openExternalCallback = useCallback(() => {}, []);
 
-  useEffect(() => {
-    if (!settings.enableFullscreenBreak) {
-      send(SHOW_WINDOW);
-    }
-  }, [send, timer.timerType, settings.enableFullscreenBreak]);
-
-  useEffect(() => {
-    send(SET_ALWAYS_ON_TOP, {
-      alwaysOnTop: settings.alwaysOnTop,
-    });
-  }, [send, settings.alwaysOnTop]);
-
-  useEffect(() => {
-    send(SET_FULLSCREEN_BREAK, {
-      shouldFullscreen,
-      alwaysOnTop: settings.alwaysOnTop,
-    });
-  }, [send, settings.alwaysOnTop, shouldFullscreen]);
-
-  useEffect(() => {
-    send(SET_COMPACT_MODE, {
-      compactMode: settings.compactMode,
-    });
-  }, [send, settings.compactMode]);
-
-  useEffect(() => {
-    send(SET_UI_THEME, {
-      isDarkMode: settings.enableDarkTheme,
-    });
-  }, [send, settings.enableDarkTheme]);
-
-  useEffect(() => {
-    send(SET_NATIVE_TITLEBAR, {
-      useNativeTitlebar: settings.useNativeTitlebar,
-    });
-  }, [send, settings.useNativeTitlebar]);
-
-  useTrayIconUpdates((dataUrl) => {
-    send(TRAY_ICON_UPDATE, { dataUrl });
-  });
-
   // Workaround to make sure it only calls once on mount
   const checkUpdate = useCallback(() => {
-    send(CHECK_FOR_UPDATES, {
+    TauriInvoker.send(CHECK_FOR_UPDATES, {
       ignoreVersion: settings.ignoreUpdate || "",
     });
-  }, [send, settings.ignoreUpdate]);
+  }, [settings.ignoreUpdate]);
 
   useEffect(() => {
     checkUpdate();
